+++ functional-tests/functional/swift/main.swift
@@ -86,7 +86,6 @@ func getAllTests() -> [XCTestCaseEntry] {
         testCase(StructsWithMethodsTests.allTests),
         testCase(StructWithInstanceTests.allTests),
         testCase(TypeDefTests.allTests),
-        testCase(VisibilityPrivateTests.allTests),
         testCase(WeakListenersTests.allTests)
     ]
     return allTests