+++ gluecodium/src/main/java/com/here/gluecodium/generator/swift/SwiftNameResolver.kt
@@ -90,7 +90,6 @@ internal class SwiftNameResolver(
         when (limeVisibility) {
             LimeVisibility.PUBLIC -> "public"
             LimeVisibility.INTERNAL -> "internal"
-            LimeVisibility.PRIVATE -> "fileprivate"
         }
 
     private fun resolveBasicType(typeId: TypeId) =